move "different" function to newly created utils.ts file
