+++ src/actions/assign.ts
@@ -1,47 +1,50 @@
-import { Context } from 'probot'
+import { Context } from "probot"
+import { different } from "../utils"
 
-export = async function assign (context: Context) {
-  // Get PR number and current PR labels
+export = async function assign(context: Context) {
+  // Get PR number and current PR reviewers
   const PRNumber = context.payload.number
-  const oldReviewers: string[] = context.payload.pull_request.requested_reviewers.map((reviewer: { login: string }) => reviewer)
+  const oldReviewers: string[] = context.payload.pull_request.requested_reviewers.map(
+    (reviewer: { login: string }) => reviewer
+  )
 
   // A little bit of helpful logging
-  context.log.info('PR Number:', PRNumber)
+  context.log.info("PR Number:", PRNumber)
 
   // Get clone URL of repository and repository directory
   // Export them to the environment for the triage script
   const repoURL = context.payload.repository.clone_url
-  process.env['REPO_URL'] = repoURL
-  process.env['REPO_DIR'] = `${process.cwd()}/github`
+  process.env["REPO_URL"] = repoURL
+  process.env["REPO_DIR"] = `${process.cwd()}/github`
 
   let newReviewers: string[] = []
   try {
     newReviewers = reviewersOfPR(PRNumber)
   } catch (err) {
-    context.log.error(err, 'Error while calling reviewersOfPR function')
+    context.log.error(err, "Error while calling reviewersOfPR function")
     return
   }
 
-  // Check if old and new labels differ. Otherwise we can skip an API call
+  // Check if old and new reviewers differ. Otherwise we can skip an API call
   if (different(oldReviewers, newReviewers)) {
-    // Log label replacement
-    context.log.info('Old reviewers:', oldReviewers)
-    context.log.info('New reviewers:', newReviewers)
+    // Log reviewers replacement
+    context.log.info("Old reviewers:", oldReviewers)
+    context.log.info("New reviewers:", newReviewers)
 
     const params = context.repo({ pull_number: PRNumber, reviewers: newReviewers })
     await context.github.pulls.createReviewRequest(params)
   } else {
-    context.log.info('No labels changed')
+    context.log.info("No reviewers changed")
   }
 }
 
 /**
- * Executes the zsh script that assign labels to the PR of PRNumber.
- * The zsh script assign the labels basing on the files that have been changed.
+ * Executes the zsh script that assign reviewers to the PR of PRNumber.
+ * The zsh script assign the reviewers basing on the files that have been changed.
  *
  * @return The list of maintainers of the changed files
  */
-function reviewersOfPR (PRNumber: number): string[] {
+function reviewersOfPR(PRNumber: number): string[] {
   // Path to script based on app root
   const zshScript = `${process.cwd()}/src/actions/reviewers.zsh`
 
@@ -53,17 +56,4 @@ function reviewersOfPR (PRNumber: number): string[] {
   // TODO: Use a ZSH script or write the logic in TypeScript
 
   return reviewers
-}
-
-function different (A: string[], B: string[]): boolean {
-  if (A.length !== B.length) return true
-
-  A.sort()
-  B.sort()
-
-  for (let i = 0; i < A.length; i++) {
-    if (A[i] !== B[i]) return true
-  }
-
-  return false
 }
